refactor: switch to functional components
